New workflow when creating a new object
This will change the workflow when a new object is added. From now on
there should be always a look be added too. Therefore the UX/UI Team
decided to implement this two step dialog:
 1. Choose a new look from Paintroid (if installed), Gallery or Camera
 2. A preview of the selected look will be given and the user can enter
the object's name
 3. If everything goes well the new object will be created and the app
will jump into the new object's script activity

Additionally if the user adds the new object within the script activity
(PointToBrick) the user will be asked to jump to the new object.